cc soumith yanboliang penguinwu anijain2305 EikanWang jgong5 Guobing-Chen XiaobingSuper zhuhaozhe blzheng Xia-Weiwen wenzhe-nrv jiayisunx peterbell10 desertfire

[ghstack-poisoned]

+++ torch/_inductor/codecache.py
@@ -63,8 +63,14 @@ def cache_dir():
 
 
 def remove_cache_dir():
-    import shutil
-    shutil.rmtree(cache_dir())
+    """
+    Removes the directory added automatically by inductor during compilation.
+    Uses the cache_dir function above.
+
+    No op if the directory does not exist.
+    """
+    if os.path.isdir(cache_dir()):
+        shutil.rmtree(cache_dir())
 
 
 class DiskCache:
